+++ Headers/Odometrie_HL.h
@@ -3,5 +3,5 @@
 void Gestion_Odometrie_HL(signed long Delta_Tick_droit,signed long Delta_Tick_gauche);
 extern signed long Odo_pos_x;
 extern signed long Odo_pos_y;
-extern long double Odo_angle;
+extern double Odo_angle;
 #endif /* ODOMETRIE_HL_H_ */
